requirements.txt
Django==4.1.5
google-api-python-client==2.74.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0